- added error middleware - fixed lint errors
