+++ graphql-gateway/src/main.js
@@ -15,6 +15,7 @@ import LogMiddleware from './middlewares/log.middleware'
 import ValidationMiddleware from './middlewares/validation.middleware'
 import HookMiddleware from './middlewares/hook.middleware'
 import AuthenticationMiddleware from './middlewares/authentication.middleware'
+import ErrorMiddleware from './middlewares/error.middleware'
 
 import logger from './logger'
 import Server from './server'
@@ -98,7 +99,7 @@ const main = async () => {
       pubsub,
       logger
     },
-    [LogMiddleware, AuthenticationMiddleware, ValidationMiddleware, HookMiddleware]
+    [ErrorMiddleware, LogMiddleware, AuthenticationMiddleware, ValidationMiddleware, HookMiddleware]
   )
 
   const httpServer = await server.start(
