+++ graphql-gateway/src/resolvers/posts/my-posts.query.js
@@ -1,7 +1,7 @@
 const myPosts = {
   authRequired: true,
-  beforeResolve: async (parent, args, {request}) => {
-    let query = { where: {author: args.user} }
+  beforeResolve: async (parent, args, { request }) => {
+    let query = { where: { author: args.user } }
 
     if (args.query) query = { where: { author: args.user, title: { $like: args.query } } }
